pyproject.toml
[project]
name = "schedulr-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.6.1",
    "fastapi>=0.116.1",
    "fastapi-pagination>=0.13.3",
    "google-api-python-client>=2.178.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langgraph>=0.6.3",
    "langgraph-checkpoint-postgres>=2.0.23",
    "psycopg[binary,pool]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "pytz>=2025.2",
    "sqlalchemy[asyncio]>=2.0.42",
    "sqlmodel>=0.0.24",
    "streamlit>=1.48.0",
    "tzlocal>=5.3.1",
    "uvicorn[standard]>=0.35.0",
]
